\name{altmetrics.analysis-package}
\alias{altmetrics.analysis-package}
\docType{package}
\title{altmetrics analysis for PLoS Study}
\description{Scripts for cleaning and processing altmetrics data for PLoS Study}
\details{
\tabular{ll}{Package: \tab altmetrics.analysis\cr
Title: \tab altmetrics analysis for PLoS Study\cr
Type: \tab Package\cr
Version: \tab 0.8\cr
Date: \tab 2011-05-05\cr
Author: \tab Jason Priem and Heather Piwowar\cr
Maintainer: \tab Jason Priem <jason@jasonpriem.com>\cr
URL: \tab http://github.com/jasonpriem/plos_altmetrics_study/wiki\cr
License: \tab MIT\cr
LazyLoad: \tab yes\cr
Depends: \tab methods\cr}
}
\author{Jason Priem <jason@jasonpriem.com>
Jason Priem and Heather Piwowar}

\keyword{ package }


