\name{applyWeightedWindow}
\alias{applyWeightedWindow}
\title{applyWeightedWindow}
\description{Weight the values of a signal within a given time window}
\usage{
applyWeightedWindow(i, x, y)
}
\arguments{
  \item{i}{index along the time and signal vectors}
  \item{x}{vector of time values}
  \item{y}{vector of signal amplitudes}
}
\details{Uses the rms and signal libraries}
\value{Return the convolved signal}

\author{Jason Priem and Heather Piwowar}





