\name{calc.correlations}
\alias{calc.correlations}
\title{calc correlations}
\description{Calculates correlations of matrix rows and columns, 
Code inspired by hetcore.data.frame from library(polycor)}
\usage{
calc.correlations(dat, missing.method = c("complete.obs", "pairwise.complete.obs"), correlation.method = c("pearson", "spearman"))
}
\arguments{
  \item{dat}{
}
  \item{missing.method}{
}
  \item{correlation.method}{
}
}

\value{Return the correlation matrix}

\author{Jason Priem and Heather Piwowar}





