\name{clean.raw.altmetrics.event.counts}
\alias{clean.raw.altmetrics.event.counts}
\title{clean raw altmetrics event counts}
\description{cleans raw altmetrics event count data}
\usage{
clean.raw.altmetrics.event.counts(dat.raw.eventcounts)
}
\arguments{
  \item{dat.raw.eventcounts}{filename of raw data, in .txt or .txt.gz format}
}



\author{Jason Priem and Heather Piwowar}





