\name{corr_pearson_normalized_create}
\alias{corr_pearson_normalized_create}
\title{corr pearson normalized create}

\usage{
corr_pearson_normalized_create(dat.norm)
}
\arguments{
  \item{dat.norm}{Data frame to for correlation}
}

\value{Returning the correlation}

\author{Jason Priem and Heather Piwowar}





