\name{event_counts_factor_scores_create}
\alias{event_counts_factor_scores_create}
\title{event counts factor scores create}

\usage{
event_counts_factor_scores_create(mycorr, dat, num.factors = 6)
}
\arguments{
  \item{mycorr}{
}
  \item{dat}{
}
  \item{num.factors}{
}
}

\value{Returning some things}

\author{Jason Priem and Heather Piwowar}





