\name{event_counts_merged_cleaned_create}
\alias{event_counts_merged_cleaned_create}
\title{event counts merged cleaned create}
\description{Merges ISI Web of Science eventcounts into other altmetrics eventcounts by DOI}
\usage{
event_counts_merged_cleaned_create(dat.eventcounts.nowos, dat.eventcounts.wos)
}
\arguments{
  \item{dat.eventcounts.nowos}{data frame with eventcounts without WoS}
  \item{dat.eventcounts.wos}{data frame with WoS eventcounts}
}

\value{Data frame with altmetrics and ISI Web of Science eventcounts}

\author{Jason Priem and Heather Piwowar}





