\name{event_counts_normalized_create}
\alias{event_counts_normalized_create}
\title{event counts normalized create}

\usage{
event_counts_normalized_create(dat.eventcounts)
}
\arguments{
  \item{dat.eventcounts}{
}
}

\value{Returning a data frame with normalized data}

\author{Jason Priem and Heather Piwowar}





