\name{event_counts_preprocessing}
\alias{event_counts_preprocessing}
\title{event counts preprocessing}
\description{Do all the preprocessing on the altmetrics and ISI WoS data for PLoS study}
\usage{
event_counts_preprocessing()
}

\value{Returns the filename of the final somethingorother}

\author{Jason Priem and Heather Piwowar}
\note{documented using inlinedocs at http://inlinedocs.r-forge.r-project.org/

read.csv works better with .gz than zip, so working around it for now}




