\name{event_counts_research_create}
\alias{event_counts_research_create}
\title{event counts research create}
\description{Filters to just research articles}
\usage{
event_counts_research_create(dat.eventcounts)
}
\arguments{
  \item{dat.eventcounts}{data frame with eventcounts}
}

\value{Return data frame with just research articles}

\author{Jason Priem and Heather Piwowar}





