\name{events_raw_wos_2008_create}
\alias{events_raw_wos_2008_create}
\title{events raw wos 2008 create}
\description{Merges ISI Web of Science eventcounts into other altmetrics eventcounts by DOI}
\usage{
events_raw_wos_2008_create(dat.raw.wos)
}
\arguments{
  \item{dat.raw.wos}{Exracted WoS event data}
}



\author{Jason Priem and Heather Piwowar}





