\name{factor.scores.bartlett}
\alias{factor.scores.bartlett}
\title{factor scores bartlett}
\description{Calculate factor scores using Bartlett method
Code is based on the middle of the factanal function in the stats package}
\usage{
factor.scores.bartlett(x, fa.fit, na.action = NULL)
}
\arguments{
  \item{x}{a dataframe}
  \item{fa.fit}{the factor analysis results; the output of calling fa()}
  \item{na.action}{what to do about NAs}
}

\value{Return the factor scores}

\author{Jason Priem and Heather Piwowar}





