\name{plot_background}
\alias{plot_background}
\title{plot background}

\usage{
plot_background(dat.input, dat.background, cols, title, xrange, yrange, colour)
}
\arguments{
  \item{dat.input}{
}
  \item{dat.background}{
}
  \item{cols}{
}
  \item{title}{
}
  \item{xrange}{
}
  \item{yrange}{
}
  \item{colour}{
}
}



\author{Jason Priem and Heather Piwowar}





