\name{plot_the_background2}
\alias{plot_the_background2}
\title{plot the background2}

\usage{
plot_the_background2(dat, dat.background, journals, column.names.to.normalize)
}
\arguments{
  \item{dat}{
}
  \item{dat.background}{
}
  \item{journals}{
}
  \item{column.names.to.normalize}{
}
}



\author{Jason Priem and Heather Piwowar}





