\name{research_articles_only}
\alias{research_articles_only}
\title{research articles only}
\description{Filters article data to include only those with articleType == "Research Article"}
\usage{
research_articles_only(dat)
}
\arguments{
  \item{dat}{data frame with eventcounts}
}

\value{Return data frame with just research articles}

\author{Jason Priem and Heather Piwowar}





